+++ tmc3/AttributeEncoder.cpp
@@ -664,11 +664,11 @@ AttributeEncoder::encodeReflectancesTransformRaht(
     int x = int(position[0]);
     int y = int(position[1]);
     int z = int(position[2]);
-    long long mortonCode = 0;
+    uint64_t mortonCode = 0;
     for (int b = 0; b < aps.raht_depth; b++) {
-      mortonCode |= (long long)((x >> b) & 1) << (3 * b + 2);
-      mortonCode |= (long long)((y >> b) & 1) << (3 * b + 1);
-      mortonCode |= (long long)((z >> b) & 1) << (3 * b);
+      mortonCode |= (uint64_t)((x >> b) & 1) << (3 * b + 2);
+      mortonCode |= (uint64_t)((y >> b) & 1) << (3 * b + 1);
+      mortonCode |= (uint64_t)((z >> b) & 1) << (3 * b);
     }
     packedVoxel[n].mortonCode = mortonCode;
     packedVoxel[n].index = n;
@@ -676,7 +676,7 @@ AttributeEncoder::encodeReflectancesTransformRaht(
   sort(packedVoxel.begin(), packedVoxel.end());
 
   // Allocate arrays.
-  long long* mortonCode = new long long[voxelCount];
+  uint64_t* mortonCode = new uint64_t[voxelCount];
   const int attribCount = 1;
   FixedPoint* attributes = new FixedPoint[attribCount * voxelCount];
   int* integerizedAttributes = new int[attribCount * voxelCount];
@@ -749,11 +749,11 @@ AttributeEncoder::encodeColorsTransformRaht(
     int x = int(position[0]);
     int y = int(position[1]);
     int z = int(position[2]);
-    long long mortonCode = 0;
+    uint64_t mortonCode = 0;
     for (int b = 0; b < aps.raht_depth; b++) {
-      mortonCode |= (long long)((x >> b) & 1) << (3 * b + 2);
-      mortonCode |= (long long)((y >> b) & 1) << (3 * b + 1);
-      mortonCode |= (long long)((z >> b) & 1) << (3 * b);
+      mortonCode |= (uint64_t)((x >> b) & 1) << (3 * b + 2);
+      mortonCode |= (uint64_t)((y >> b) & 1) << (3 * b + 1);
+      mortonCode |= (uint64_t)((z >> b) & 1) << (3 * b);
     }
     packedVoxel[n].mortonCode = mortonCode;
     packedVoxel[n].index = n;
@@ -761,7 +761,7 @@ AttributeEncoder::encodeColorsTransformRaht(
   sort(packedVoxel.begin(), packedVoxel.end());
 
   // Allocate arrays.
-  long long* mortonCode = new long long[voxelCount];
+  uint64_t* mortonCode = new uint64_t[voxelCount];
   const int attribCount = 3;
   FixedPoint* attributes = new FixedPoint[attribCount * voxelCount];
   int* integerizedAttributes = new int[attribCount * voxelCount];
