+++ tmc3/RAHT.cpp
@@ -136,7 +136,7 @@ rahtFixedPointInverseRotation(
 void
 regionAdaptiveHierarchicalTransform(
   FixedPoint quantStepSizeLuma,
-  long long* mortonCode,
+  uint64_t* mortonCode,
   FixedPoint* attributes,
   uint64_t* weight,
   int* binaryLayer,
@@ -151,7 +151,7 @@ regionAdaptiveHierarchicalTransform(
 
   FixedPoint* attributesTransformed = new FixedPoint[voxelCount * attribCount];
   uint64_t* weightTransformed = new uint64_t[voxelCount];
-  long long* mortonCodeTransformed = new long long[voxelCount];
+  uint64_t* mortonCodeTransformed = new uint64_t[voxelCount];
 
   d = 0;
   while (N > 1 || d % 2) {
@@ -235,7 +235,7 @@ regionAdaptiveHierarchicalTransform(
 void
 regionAdaptiveHierarchicalInverseTransform(
   FixedPoint quantStepSizeLuma,
-  long long* mortonCode,
+  uint64_t* mortonCode,
   FixedPoint* attributes,
   uint64_t* weight,
   const int attribCount,
@@ -249,21 +249,21 @@ regionAdaptiveHierarchicalInverseTransform(
 
   FixedPoint* attributesTransformed = new FixedPoint[voxelCount * attribCount];
   uint64_t* weightTransformed = new uint64_t[voxelCount];
-  long long* mortonCodeTransformed = new long long[voxelCount];
+  uint64_t* mortonCodeTransformed = new uint64_t[voxelCount];
 
-  long long** mortonCodeBuffer;
+  uint64_t** mortonCodeBuffer;
   uint64_t** weightBuffer;
   size_t* M_buffer;
   {
     size_t depth = 0;
-    long long maxMortonCode = mortonCode[voxelCount - 1];
+    uint64_t maxMortonCode = mortonCode[voxelCount - 1];
 
     while (maxMortonCode--) {
       depth++;
       maxMortonCode >>= 1;
     }
 
-    mortonCodeBuffer = new long long*[depth + 1];
+    mortonCodeBuffer = new uint64_t*[depth + 1];
     weightBuffer = new uint64_t*[depth + 1];
     M_buffer = new size_t[depth + 1];
   }
@@ -285,7 +285,7 @@ regionAdaptiveHierarchicalInverseTransform(
   // Re-obtain weights at the decoder by partially executing the encoder
   d = 0;
   while (N > 1) {
-    long long* _mortonCode = new long long[M];
+    uint64_t* _mortonCode = new uint64_t[M];
     uint64_t* _weight = new uint64_t[M];
 
     M_buffer[d] = M;
